requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
arrow==1.2.3
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
binaryornot==0.4.4
certifi==2023.5.7
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cookiecutter==2.1.1
Django==4.2.2
django-browser-reload==1.10.0
django-tailwind==3.5.0
django-tailwindcss==0.4.0
frozenlist==1.3.3
idna==3.4
Jinja2==3.1.2
jinja2-time==0.2.0
MarkupSafe==2.1.3
multidict==6.0.4
openai==0.27.8
Pillow==9.5.0
psycopg2-binary==2.9.6
python-dateutil==2.8.2
python-dotenv==1.0.0
python-slugify==8.0.1
PyYAML==6.0
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
text-unidecode==1.3
tqdm==4.65.0
typing_extensions==4.6.3
tzdata==2023.3
urllib3==2.0.3
yarl==1.9.2
gunicorn